client/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
grpcio==1.58.0
grpcio-tools==1.58.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
protobuf==4.24.3
requests==2.31.0
urllib3==2.0.4
Werkzeug==2.3.7